Make crate `no-std` compatible
Copying the pattern from the `lightning` crate, we re-export
`alloc`/`hashbrown` or `std` depending on whether the `std` feature is
set. We also copy some `sync` types from the `lightning` crate that
are `no-std` compatible.

Moreover, we check no-std compatibility in CI and, since we already
touched most of them, reorder and cleanup the `use` statements in all
modules.