Add missing docs for `lsps1/msgs.rs`, make submodules `pub`
In parallel to LSPS2, we expose the internal
`msgs`/`message_handler`/`event` modules of LSPS1.

To this end, we add docs to all message types, mark docs in `event.rs`
as TODO, and hide the public methods of LSPS1's message handler, as
we'll want to expose-and-document them individually at a later date.