Drop `MessageQueue` trait and rename `DefaultMessageQueue`
We previously introduced the `MessageQueue` trait in order to be able to
move the `APeerManager` dependency to a single (optional) place. As we
now got rid of the type dependency, we may as well drop this
intermediary step again, as there is likely no immediate benefit in
allowing users to implement their own custom message queues. However,
removing the type parameter further cleans up our types and reduces
overall complexity.