Ignore rather than disconnect peers who send us bogus data
Disconnecting might be unfortunate if we have open channels with the
counterparty. If they send us bogus data, we now just add them to a
(currently unpersisted) ignore list.

We should in the future figure out when to drop them from this list, so
that peers have the chance to recover.